Adds ability to edit and delete posts
